gooty_proxy::inspection::cidr

Struct Cidr

pub struct Cidr {
    pub network_address: IpAddr,
    pub prefix_length: u8,
    pub cidr_string: String,
}

Represents a CIDR (Classless Inter-Domain Routing) block.

This struct provides functionality for working with IP subnet ranges, including checking if IPs are contained within a block.

Examples

use gooty_proxy::inspection::Cidr;
use std::net::{IpAddr, Ipv4Addr};

let cidr = Cidr::to_cidr("192.168.1.0/24").unwrap();
let ip = IpAddr::V4(Ipv4Addr::new(192, 168, 1, 10));
assert!(cidr.contains(&ip));

Fields
network_address: IpAddr

Base IP address of the network

prefix_length: u8

Network prefix length (subnet mask bits)

cidr_string: String

String representation of the CIDR

Implementations

impl Cidr

pub fn to_cidr(cidr_str: &str) -> CidrResult<Self>

Creates a new CIDR from a string representation (e.g., โ192.168.0.0/24โ).

Arguments

cidr_str - A string in CIDR notation format like โ192.168.0.0/24โ

Returns

A CidrResult<Cidr> which is Ok if parsing succeeded or an error if the CIDR notation is invalid.

Errors

This function will return an error if:

The format is invalid (not in format โip/prefixโ)

The IP address part cannot be parsed

The prefix length is invalid (>32 for IPv4 or >128 for IPv6)

Examples

use gooty_proxy::inspection::Cidr;

let cidr = Cidr::to_cidr("10.0.0.0/8").unwrap();
assert_eq!(cidr.get_prefix_length(), 8);

pub fn contains(&self, ip: &IpAddr) -> bool

Checks if an IP address is contained within this CIDR block.

This method compares the network bits of the provided IP with the network bits of this CIDR block to determine if the IP is within the subnet.

Arguments

ip - The IP address to check

Returns

true if the IP is within this CIDR block, false otherwise.

Examples

use gooty_proxy::inspection::Cidr;
use std::net::{IpAddr, Ipv4Addr};

let cidr = Cidr::to_cidr("192.168.1.0/24").unwrap();
let ip_in = IpAddr::V4(Ipv4Addr::new(192, 168, 1, 10));
let ip_out = IpAddr::V4(Ipv4Addr::new(192, 168, 2, 10));

assert!(cidr.contains(&ip_in));
assert!(!cidr.contains(&ip_out));

pub fn get_network_address(&self) -> &IpAddr

Returns the network address of the CIDR block.

Returns

A reference to the network IP address.

pub fn get_prefix_length(&self) -> u8

Returns the prefix length of the CIDR block.

Returns

The prefix length as a u8 value (e.g., 24 for a /24 network).

pub fn to_string(&self) -> &str

Returns the string representation of the CIDR block.

Returns

The CIDR in string format like โ192.168.1.0/24โ.