gooty_proxy::orchestration::threading

Function execute_with_concurrency_limit

pub async fn execute_with_concurrency_limit<T, F, Fut>(
    items: Vec<T>,
    concurrency: usize,
    job_fn: F,
) -> Vec<Pin<Box<dyn Future<Output = ()> + Send>>>
where
    T: Send + 'static,
    F: FnMut(T) -> Fut + Send,
    Fut: Future<Output = ()> + Send + 'static,

Execute multiple futures concurrently with a limit on parallelism

Panics

This function will panic if the semaphore is closed, which can happen if the semaphore is dropped while permits are still active.