gooty_proxy::definitions::proxy

Struct Proxy

pub struct Proxy {

    pub proxy_type: ProxyType,
    pub address: IpAddr,
    pub port: u16,
    pub username: Option<String>,
    pub password: Option<String>,
    pub anonymity: AnonymityLevel,
    pub country: Option<String>,
    pub organization: Option<String>,
    pub asn: Option<String>,
    pub hostname: Option<String>,
    pub latency_ms: Option<u32>,
    pub added_at: DateTime<Utc>,
    pub last_checked_at: Option<DateTime<Utc>>,
    pub check_count: usize,
    pub check_failure_count: usize,
    pub last_used_at: Option<DateTime<Utc>>,
    pub use_count: usize,
    pub use_failure_count: usize,
    pub ip_metadata: Option<IpMetadata>,
    pub cidr: Option<String>,
    pub location: Option<Location>,
    pub network: Option<NetworkInfo>,
    pub organization_info: Option<Organization>,

}

Represents a proxy server with its connection details and metadata.

This struct is used throughout the application to manage and interact with proxy servers. It includes fields for the proxy’s type, address, port, and anonymity level, as well as methods for managing its state and statistics.

Examples

use gooty_proxy::definitions::Proxy;
use gooty_proxy::definitions::enums::{ProxyType, AnonymityLevel};
use std::net::{IpAddr, Ipv4Addr};

let proxy = Proxy::new(
    ProxyType::Http,
    IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)),
    8080,
    AnonymityLevel::Elite,
);

assert_eq!(proxy.proxy_type, ProxyType::Http);
assert_eq!(proxy.port, 8080);

Fields
proxy_type: ProxyType

The type of the proxy (e.g., HTTP, HTTPS, SOCKS4, SOCKS5).

address: IpAddr

The IP address of the proxy server.

port: u16

The port number of the proxy server.

username: Option<String>

Optional username for authentication.

password: Option<String>

Optional password for authentication.

anonymity: AnonymityLevel

The anonymity level of the proxy.

country: Option<String>

The country associated with the proxy, if available.

organization: Option<String>

The organization associated with the proxy, if available.

asn: Option<String>

The ASN (Autonomous System Number) of the proxy, if available.

hostname: Option<String>

The hostname of the proxy, if available.

latency_ms: Option<u32>

The latency of the proxy in milliseconds, if measured.

added_at: DateTime<Utc>

When the proxy was added to the system.

last_checked_at: Option<DateTime<Utc>>

When the proxy was last checked for availability.

check_count: usize

The total number of checks performed on the proxy.

check_failure_count: usize

The number of failed checks for the proxy.

last_used_at: Option<DateTime<Utc>>

When the proxy was last used for a connection.

use_count: usize

Number of times the proxy has been used for connections.

use_failure_count: usize

Number of times connections through this proxy have failed.

ip_metadata: Option<IpMetadata>

Extended network metadata for the proxy IP address.

cidr: Option<String>

CIDR notation for the network the proxy belongs to.

location: Option<Location>

Optional location information for the proxy IP address.

network: Option<NetworkInfo>

Optional network information for the proxy IP address.

organization_info: Option<Organization>

Optional organization information for the proxy IP address.

Implementations

impl Proxy

pub fn new( proxy_type: ProxyType, address: IpAddr, port: u16, anonymity: AnonymityLevel, ) -> Self

Creates a new proxy with mandatory fields and default values for statistics.

Arguments

proxy_type - The type of proxy protocol to use (HTTP, HTTPS, SOCKS4, SOCKS5)

address - The IP address of the proxy server

port - The port number the proxy server listens on

anonymity - The level of anonymity provided by the proxy

Returns

A new Proxy instance with default values for non-specified fields

Examples

use spiderling_proxy::definitions::{
    enums::{AnonymityLevel, ProxyType},
    proxy::Proxy,
};
use std::net::{IpAddr, Ipv4Addr};

let proxy = Proxy::new(
    ProxyType::Http,
    IpAddr::V4(Ipv4Addr::new(192, 168, 1, 1)),
    8080,
    AnonymityLevel::Anonymous,
);

pub fn with_auth(self, username: String, password: String) -> Self

Sets authentication credentials for the proxy.

Arguments

username - Username for proxy authentication

password - Password for proxy authentication

Returns

Self with authentication credentials set

Examples

let proxy = Proxy::new(
    ProxyType::Http,
    IpAddr::V4(Ipv4Addr::new(192, 168, 1, 1)),
    8080,
    AnonymityLevel::Anonymous
).with_auth("username".to_string(), "password".to_string());

pub fn with_country(self, country: String) -> Self

Sets the country for the proxy.

Arguments

country - The country where the proxy server is located

Returns

Self with country information set

pub fn with_hostname(self, hostname: String) -> Self

Sets the hostname for the proxy.

Arguments

hostname - The hostname of the proxy server

Returns

Self with hostname information set

pub fn with_organization(self, organization: String) -> Self

Sets the organization for the proxy.

Arguments

organization - The organization or ISP operating the proxy

Returns

Self with organization information set

pub fn validate(&self) -> Result<(), ProxyError>

Validates that the proxy configuration is correct.

Returns

Ok(()) - If the proxy configuration is valid

Err(ProxyError) - If the proxy configuration is invalid

Errors

This function will return an error if:

The port is set to 0

Authentication is missing required fields (e.g., password is missing when username is provided for SOCKS5)

pub fn record_check(&mut self, latency: u32)

Records a successful check of the proxy

pub fn record_check_failure(&mut self)

Records a failed check of the proxy

pub fn record_use(&mut self)

Records a successful use of the proxy

pub fn record_use_failure(&mut self)

Records a failed use of the proxy

pub fn check_success_rate(&self) -> f64

Calculates the success rate of the proxy based on check history

pub fn use_success_rate(&self) -> f64

Calculates the success rate of the proxy based on usage history

pub fn to_connection_string(&self) -> String

Returns a connection string representation of the proxy

pub fn update_metadata( &mut self, country: Option<String>, organization: Option<String>, hostname: Option<String>, anonymity: Option<AnonymityLevel>, )

Updates the proxy with new information from a check

pub fn update_with_ip_metadata(&mut self, metadata: IpMetadata)

Updates the proxy with network metadata from a sleuth lookup

pub fn get_ip_metadata(&self) -> Option<&IpMetadata>

Gets the full IP metadata if available

impl Proxy

Helper functions for serialization and deserialization

pub fn to_json(&self) -> Result<String, Error>

Serializes the proxy to a JSON string

pub fn from_json(json: &str) -> Result<Self, Error>

Deserializes a proxy from a JSON string